code/requirements.txt
google-cloud-bigquery
google-cloud-firestore
google-cloud-pubsub
openai
requests